# Optimization: Revert qfrc_actuator to previously _qfrc_actuator_sparse